Improved GameLoop. See description
- Increased drop rate from the following items:
AssaultRifle from 5% to 10%.
Shotgun from 5% to 10%.
Medkit increased from 5% to 20%.
Ammobox increased from 5% to 20%.
- Zombies when shot, will become agressive and will start chasing the nearest player.
- Created spawners that spawns Idle Zombies. (Basic, Fast, Boss)
- Breakable obstacles now have a NavMeshObstacle. Previously, all entities were able to pass through these obstacles.
