+++ Assets/Resources/Prefabs/Gameplay/Door.prefab
@@ -14,6 +14,7 @@ GameObject:
   - component: {fileID: 2134213903203002453}
   - component: {fileID: 977705018095458381}
   - component: {fileID: 3469008443589854374}
+  - component: {fileID: 6786303124723748192}
   m_Layer: 0
   m_Name: Door
   m_TagString: Untagged
@@ -151,6 +152,22 @@ BoxCollider:
   serializedVersion: 3
   m_Size: {x: 1.5, y: 1, z: 1}
   m_Center: {x: -0.6, y: 0.5, z: -0.7}
+--- !u!208 &6786303124723748192
+NavMeshObstacle:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7377907689856257518}
+  m_Enabled: 1
+  serializedVersion: 3
+  m_Shape: 1
+  m_Extents: {x: 0.75, y: 0.5, z: 0.5}
+  m_MoveThreshold: 0.1
+  m_Carve: 0
+  m_CarveOnlyStationary: 1
+  m_Center: {x: -0.6, y: 0.5, z: -0.7}
+  m_TimeToStationary: 0.5
 --- !u!1001 &4663273407313120689
 PrefabInstance:
   m_ObjectHideFlags: 0
