+++ Assets/Resources/Prefabs/Gameplay/Entities/Box.prefab
@@ -13,6 +13,7 @@ GameObject:
   - component: {fileID: 4590334083199595721}
   - component: {fileID: 2320510854949627292}
   - component: {fileID: 1337583059908058407}
+  - component: {fileID: 2690864453716352465}
   m_Layer: 0
   m_Name: Box
   m_TagString: Untagged
@@ -124,6 +125,22 @@ BoxCollider:
   serializedVersion: 3
   m_Size: {x: 1.28, y: 1.28, z: 1.28}
   m_Center: {x: 0, y: 0.64, z: 0}
+--- !u!208 &2690864453716352465
+NavMeshObstacle:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5983385692335723542}
+  m_Enabled: 1
+  serializedVersion: 3
+  m_Shape: 1
+  m_Extents: {x: 0.64, y: 0.64, z: 0.64}
+  m_MoveThreshold: 0.1
+  m_Carve: 0
+  m_CarveOnlyStationary: 1
+  m_Center: {x: 0, y: 0.64, z: 0}
+  m_TimeToStationary: 0.5
 --- !u!1001 &2149013360711192587
 PrefabInstance:
   m_ObjectHideFlags: 0
