+++ Assets/Resources/Prefabs/Gameplay/Entities/Car.prefab
@@ -13,6 +13,7 @@ GameObject:
   - component: {fileID: 34753933554102182}
   - component: {fileID: 8220997127095456998}
   - component: {fileID: 1192060505109604229}
+  - component: {fileID: 3477625283651442487}
   m_Layer: 0
   m_Name: Car
   m_TagString: Untagged
@@ -127,6 +128,22 @@ BoxCollider:
   serializedVersion: 3
   m_Size: {x: 2, y: 1, z: 5}
   m_Center: {x: 0, y: 0.75, z: 0}
+--- !u!208 &3477625283651442487
+NavMeshObstacle:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5257529110699039678}
+  m_Enabled: 1
+  serializedVersion: 3
+  m_Shape: 1
+  m_Extents: {x: 1, y: 0.5, z: 2.5}
+  m_MoveThreshold: 0.1
+  m_Carve: 0
+  m_CarveOnlyStationary: 1
+  m_Center: {x: 0, y: 0.75, z: 0}
+  m_TimeToStationary: 0.5
 --- !u!1001 &6319259077651420851
 PrefabInstance:
   m_ObjectHideFlags: 0
