+++ Assets/Resources/Prefabs/Gameplay/Entities/ZombieHordeSpawners/ZombieHordeSpawner_Fast.prefab.meta
