+++ Assets/Resources/Prefabs/Gameplay/Entities/ZombieIdleSpawners.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9b94a3d138ba0f543a56531f1d31f391
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
