+++ Assets/Resources/Prefabs/Gameplay/Entities/ZombieIdleSpawners/ZombieIdleSpawner_Basic.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 123e25cb226097040904fd4b61dcef8c
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
