+++ Assets/Resources/Prefabs/Gameplay/Entities/ZombieIdleSpawners/ZombieIdleSpawner_Boss.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c8be1fb5d20544a49ab8976776486697
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
