+++ Assets/Resources/Prefabs/Gameplay/Entities/ZombieIdleSpawners/ZombieIdleSpawner_Fast.prefab
@@ -1,0 +1,63 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!1001 &2396216512646156469
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 876148721350643675, guid: e4aec5cbfabe9f449a86597f7e82cc70, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 876148721350643675, guid: e4aec5cbfabe9f449a86597f7e82cc70, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 876148721350643675, guid: e4aec5cbfabe9f449a86597f7e82cc70, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 876148721350643675, guid: e4aec5cbfabe9f449a86597f7e82cc70, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 876148721350643675, guid: e4aec5cbfabe9f449a86597f7e82cc70, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 876148721350643675, guid: e4aec5cbfabe9f449a86597f7e82cc70, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 876148721350643675, guid: e4aec5cbfabe9f449a86597f7e82cc70, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 876148721350643675, guid: e4aec5cbfabe9f449a86597f7e82cc70, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 876148721350643675, guid: e4aec5cbfabe9f449a86597f7e82cc70, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 876148721350643675, guid: e4aec5cbfabe9f449a86597f7e82cc70, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1674819268700206388, guid: e4aec5cbfabe9f449a86597f7e82cc70, type: 3}
+      propertyPath: m_Name
+      value: ZombieIdleSpawner_Fast
+      objectReference: {fileID: 0}
+    - target: {fileID: 1674819268700206388, guid: e4aec5cbfabe9f449a86597f7e82cc70, type: 3}
+      propertyPath: m_TagString
+      value: FastZombieSpawnpoint
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: e4aec5cbfabe9f449a86597f7e82cc70, type: 3}
