+++ Assets/Resources/Prefabs/Gameplay/Entities/ZombieIdleSpawners/ZombieIdleSpawner_Fast.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 9b485e555c00ae8488d5b56cd89562ce
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
