+++ Assets/Resources/Prefabs/Gameplay/Entities/ZombieIdleSpawners/ZombieIdleSpawner_Template.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e4aec5cbfabe9f449a86597f7e82cc70
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
